WeatherServiceTest.java
public class WeatherServiceTest {
    private WeatherService weatherService;

    @Before
    public void setUp() {
        weatherService = new WeatherService();
    }

    @Test
    public void testGetWeatherData() {
        WeatherData data = weatherService.getWeatherData("New York");
        assertNotNull(data);
        assertEquals("New York", data.getLocation());
        assertEquals(22, data.getTemperature());
        assertEquals("Partly cloudy", data.getForecast());
        assertEquals(65, data.getHumidity());
    }

    @Test
    public void testCalculateForecastAccuracy() {
        WeatherData actualData = new WeatherData("New York", 22, "Partly cloudy", 65);
        WeatherData predictedData = new WeatherData("New York", 21, "Partly cloudy", 60);
        int accuracy = weatherService.calculateForecastAccuracy(actualData, predictedData);
        assertTrue(accuracy > 0);
    }

    @Test
    public void testSendWeatherNotification() {
        weatherService.sendWeatherNotification("Storm warning");
        // Just verify it doesn't throw exceptions, output is seen on console
    }

    private void assertNotNull(WeatherData data) {
        throw new UnsupportedOperationException("Not supported yet.");
    }

    private void assertEquals(int i, int temperature) {
        throw new UnsupportedOperationException("Not supported yet.");
    }

    private void assertEquals(String new_York, String location) {
        throw new UnsupportedOperationException("Not supported yet.");
    }

    private void assertTrue(boolean b) {
        throw new UnsupportedOperationException("Not supported yet.");
    }
}
